+++ hail/src/main/scala/is/hail/io/BufferSpecs.scala
@@ -1,5 +1,6 @@
 package is.hail.io
 
+import is.hail.compatibility.LZ4BlockBufferSpec
 import is.hail.rvd.AbstractRVDSpec
 import java.io._
 
@@ -112,7 +113,7 @@ trait BlockBufferSpec extends Spec {
   def buildCodeOutputBuffer(out: Code[OutputStream]): Code[OutputBlockBuffer]
 }
 
-sealed abstract class LZ4BlockBufferSpecCommon extends BlockBufferSpec {
+abstract class LZ4BlockBufferSpecCommon extends BlockBufferSpec {
   require(blockSize <= (1 << 16))
 
   def typeName: String
@@ -132,13 +133,6 @@ sealed abstract class LZ4BlockBufferSpecCommon extends BlockBufferSpec {
 
   def buildCodeOutputBuffer(out: Code[OutputStream]): Code[OutputBlockBuffer] =
     Code.newInstance[LZ4OutputBlockBuffer, Int, OutputBlockBuffer](blockSize, child.buildCodeOutputBuffer(out))
-}
-
-@deprecated("LZ4HCBlockBufferSpec is a drop-in replacement for this class", "Hail 0.2.24")
-final case class LZ4BlockBufferSpec(blockSize: Int, child: BlockBufferSpec)
-    extends LZ4BlockBufferSpecCommon {
-  def lz4 = LZ4.hc
-  def typeName = "LZ4BlockBufferSpec"
 }
 
 final case class LZ4HCBlockBufferSpec(blockSize: Int, child: BlockBufferSpec)
